requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
groq
PyPDF2
python-dotenv
pypdf
